% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_index.R
\name{xml_index}
\alias{xml_index}
\title{Index a PubMedCentral XML file for fast record access.}
\usage{
xml_index(xml_file = "pmc_result.xml", db)
}
\arguments{
\item{xml_file}{\code{character(1)} file path to the PubMedCentral XML
file.}

\item{db}{a database object returned by \code{pmcbioc_db()} with
\code{read_only = FALSE}.}
}
\value{
\code{xml_index()} returns the \code{pmcbioc_db} database argument
\code{db}.
}
\description{
\code{xml_index()} creates an index for the XML file
retrieved from PubMedCentral, storing the index in the DuckDB
database. Use \code{xml_xpath()} to query the index.
}
