% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{xml_parse}
\alias{xml_parse}
\title{Parse a PubMedCentral XML file to a DuckDB database}
\usage{
xml_parse(xml_file, db)
}
\value{
\code{xml_parse()} returns the \code{pmcbioc_db} database passed as argument
\code{db}, updated to include the tables described here. The tables are
as follows:
\itemize{
\item \code{article} includes columns \code{pmcid}, \code{title}, \code{journal}, \code{year},
and \code{pmid}.
\item \code{author} is a one-to-many map between \code{pmcid} and author
\code{surname} and \code{givenname}.
\item \code{keyword} is a one-to-many map between \code{pmcid} and \code{keyword}.
\item \code{refpmid} is a one-to-many map between \code{pmcid} and the \code{refpmid}
PubMed identifiers of cited references.
}
}
\description{
\code{xml_parse()} processes a PubMedCentral XML file by
extracting XML elements, transforming them to text or integer
representations, and loading the result into a database.
}
\details{
\code{xml_parse()} can be slow, e.g., 1000 records per minute, and grow
to consume a large amount of memory, e.g., 18 Gb.

\code{xml_parse()} uses \code{XML::xmlEventParse()} to iterate through the
XML file. Each \verb{//article} branch is queried using XPath
expressions. The expressions and subsequent transformations are
meant to extract the following information; currently, not all
records are processed correctly.
\itemize{
\item \code{pmcid}: The PubMedCentral identifier associated with the record.
\item \code{title}: The article title.
\item \code{journal}: The journal in which the article was published.
\item \code{year}: The year of publication, represented as an
integer. Articles may have several publication dates (e.g.,
electronically before physically). \code{year} is the earliest date in
the record.
\item \code{pmid}: PubMed identifier of the record.
\item \code{surname}: Surname of each author. Some authors have only a surname.
\item \code{givenname}: Given name(s) of each author. Some authors have only
given names.
\item \code{keyword}: Keywords associated with the publication. Keywords are
not standardized
\item \code{refpmid}: PubMed identifiers of each reference in the record.
}

\code{pmcid} is used as a key across database tables.
}
