% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_xpath.R
\name{xml_xpath}
\alias{xml_xpath}
\title{Use XPath to query records of a PubMedCentral XML file}
\usage{
xml_xpath(db_index, xpath = NULL, xml_file = "pmc_result.xml")
}
\arguments{
\item{`db_index`}{a tibble with columns \code{start} and \code{length},
typcially the result of filtering the \code{index} table of a
\code{pmcbioc_db()} database.}

\item{`xpath`}{\code{NULL} or a \code{character(1)} XPath expression to query
the records in \code{db_index}.}
}
\value{
When \code{xpath = NULL}, \code{xml_xpath()} returns an XML object
representing all records in \code{db_index} as a
\verb{<pmc-articleset>...</pmc-articleset>}.

When \code{xpath} is \code{character(1)}, the records in \code{db_index}
are queried using XPath. The return value can be coerced to an
\emph{R} representation with \code{XML::xmlValue()}.
}
\description{
\code{xml_xpath()} queries an indexed PubMedCentral XML
file, returning records indicated by \code{db_index} and queried
using \code{xpath}.
}
